Safe contains usage
- Fixed, some string values not being escaped correctly for contains
expression.

+++ test/Datalist.Tests/Objects/Models/NumericIdModel.cs
@@ -5,6 +5,6 @@ namespace Datalist.Tests.Objects
     public class NumericIdModel
     {
         [DatalistColumn]
-        public Decimal Id { get; set; }
+        public Int32 Id { get; set; }
     }
 }
